macos-api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psutil==5.9.8
pydantic==2.6.3
python-multipart==0.0.9
pillow==10.2.0
websocket-client==1.7.0
